fix: add missing receiptsEndpoint to W3UIProvider (#144)
## Description

This PR adds a `receiptsEndpoint` configuration to `W3UIProvider`. The
receipts URL is currently not provided by the console to the client as
an option, so this ensures the endpoint is explicitly configured.

## Motivation
Uploading in staging is currently broken. The console shows an error
when attempting to retrieve the receipt.

The receipts endpoint being used for staging is incorrect due to a
missing configuration. This causes the value to default to a hardcoded
production endpoint.

---------

Co-authored-by: ash <alan@storacha.network>