fix: gateway authorization based on env (#174)
In order to create new spaces in `Staging` or `localhost` environments
we need to configure the w3up-client to authorize the correct gateway
service, otherwise, it will default to the Production Storacha Gateway.
This PR implements the logic to create the gateway connection based on
the environment variables and passes that connection to the
`w3up-client.createSpace` call, so the correct service is authorized.