fix: do not add space to location claim for legacy provisioned spaces (#178)
### Problem

⚠️🚦 Serving content from the gateway 🚦⚠️

I've realized that we're using the presence of space in location claims
to activate access control authorization checks.

When we deploy the consolidated upload-service repos ALL location claims
will have a space - we lose the code path where we can add the property
or not, since it is in the storage node.

😱 😱 😱

### Proposal

There are two ways to store content. If your space is provisioned by
`did:web:up.storacha.network` then it goes to a storage node, if it has
been provisioned by `did:web:web3.storage` then the upload service will
store the data.

If the upload service stores the data we use the “old flow”, i.e.
`web3.storage/blob/allocate` and `web3.storage/blob/accept`. We store
the data in carpark as before and we generate the location commitment as
before (without a space). The location commitment is published to the
content-claims service as before, so that it can appear in (legacy)
results from the indexing service. It means we have to keep the content
claims service running a bit longer (writes) 😢.

The `blob/index/add` invocation will invoke the new indexer **even for
legacy space data**. This allows us to still move away from writing to
the legacy dynamodb block index.

I think this is largely in line with what we were originally
thinking…and allows us to keep space in location commitment as
indication of required content serve delegation.

### This PR

In this PR I've added back the `web3.storage/blob/*` handlers, with a
few modifications to allow them to be called directly from the
`space/blob/add` handler and the `ucan/conclude` for `http/put` handler
if we detect that the space is provisioned with `did:web:web3.storage`.