feat: allow multiple proofs (#242)
This PR updates UCN to allow multiple proofs. You might want to pass
multiple proofs in the case when you use a space DID as the name, and
want to pass an attestation in addition to the delegation.

This necessitates a method for serializing a name with _muliple_ proofs
(we only allowed one delegation before so could use the `.archive()`
method of the delegation), hence names now have an `.archive()` function
that will export blocks from all delegations into a single archive.