feat: add `app` to `access/authorize` and `access/confirm`
We need a way to ensure users are redirected back to `https://bsky.storage` after completing the Stripe checkout process.

To facilitate this, add an `app` field to the `nb` of `access/authorize` and `access/confirm` - this can be set to `bsky-backups`, which will tell the code in `w3infra` to render the same pricing table we use on https://bsky.storage which sends the user back to that app after picking a plan via the email flow.

There are a bunch of ways we could do this, but in the interest of shipping ASAP I think this is a reasonable and fast way to go.