feat: implement replication protocol (#253)
Extends https://github.com/storacha/upload-service/pull/249 to add
replication protocol implementation.

Adds a `replicaStore` and `maxReplicas` to service context.

The replica store tracks requested replicas for a given `space`+`blob`.
In addition to the space DID and blob hash, the replica store stores the
replica node that the replica was allocated on, and the known status of
the replication. The status can be `allocated`, `transferred` or
`failed`.

Invoking `space/blob/replicate` validates the blob exists in the space,
that no more than `maxReplicas` have been allocated in total and that
the location commitment for the primary is valid.

The router interface has been extended to allow it to select multiple
replica nodes that do not include the primary node.

The `ucan/conclude` handler has been extended to allow a
`blob/replica/transfer` receipt to be sent. This updates the status of
the replica in our DB (and makes the receipt available for the client to
retrieve).