fix!: legacy space claim publishing (#278)
This PR fixes the claim publishing for legacy spaces (spaces provisioned
with `did:web:web3.storage`).

Currently when uploading to a legacy space we'll try to invoke
`assert/location` on the indexing service. However the indexing service
does not support this invocation (the legacy content claims services
does). This broke when we swapped the invocation config for the content
claims service in place for the invocation config for the indexing
service. They are mostly the same, but instead of an `assert/location`
handler, the indexing service exposes an `claim/cache` handler that the
storage nodes use to cache location commitments.

BREAKING CHANGE: This PR seperates the two services in code - the
`claimsService` in context previously had _become_ the indexing service,
but now there is explicitly an `indexingService` in context _as well as_
a `claimsService`.

The `indexingService` is used by `space/index/add` to send an index
claim, and the `claimsService` is used by `web3.storage/blob/accept` to
send a location commitment to the legacy content claims service. Note:
the `claimsService` is used by `space/index/add` in the case where the
space is a legacy space.

Finally, the IPNI service that publishes IPNI info for the _public_
network now recieves the space and it's providers as parameters. i.e.
the signature of the `publish` method of `IPNIService` has changed. This
allows us to write to the legacy block index table for legacy spaces but
not for new spaces.

---------

Co-authored-by: Vicente Olmedo <vicente@storacha.network>