feat: block deduplication (#311)
This PR adds a new option `dedupe: boolean` which is `true` by default.

When `true`, the client will keep track of the CIDs of blocks that have
been uploaded and will not add blocks that have already been seen to the
uploaded CAR shards.

If your dataset contains a bunch of repeated files/data then this can
reduce upload size and increase upload speed.