fix: memory leak (#312)
This PR fixes a memory leak that occurs when a user uploads a CAR file
using the client.

The CAR reader is very efficient when reading CARs - it does not copy
memory and instead creates "views" over the data i.e. a `TypedArray`
like `Uint8Array`. We store block digests and positions during the
upload so that after everything has completed we can create and upload
an index. However by storing the digests, we are inadvertently storing
many more bytes than we care about due to the fact that the digest bytes
are actually an view over a much bigger underlying array buffer 🙈.

This issue does not occur when the data source is non-CAR (i.e. files
and directories), since new memory is allocated for each hash as it is
generated.

The PR fixes the issue by making a copy of the block digests if they are
a view over a bigger buffer. For file/directory uploads this should have
no performance impact however for CAR uploads this will have an adverse
affect performance but will prevent OOM.

resolves https://github.com/storacha/w3cli/issues/193