fix(console): origin verification for msg channels (#325)
`window.postMessage()` events: Have an origin property for security
validation - I was using that before the refactor to `MessageChannel`.
Locally, it works fine, but in Staging, it started denying the messages
from `https://staging.console.storacha.network`.

`MessagePort events`: Do NOT have an origin property because the
security is established when the MessageChannel is created and ports are
transferred. This is more secure than validating the origin because only
the validated origin has the port to send messages.