fix(access): include session proofs issued by service aliases in invocations (#328)
The client sometimes uses `agent.invokeAndExecute(...)` or
`agent.invoke(...)`, which by default scopes session proofs to the
invocation audience or `connection.id`. If the invocation audience or
`connection.id` is `did:web:web3.storage` and the session proof was
issued by `did:web:up.storacha.network` then the session proof will not
be included in the invocation and it will fail.

This PR alters the access client to scope session proofs to the
invocation audience or `connection.id` as well as any alias DIDs the
service is known by.

Due to the default values for service identities there should be no
code/config changes necessary to apps using the client - just a
dependency update.