fix: detect "plan not found" errors correctly
this used to work accidentally because we would render the plan picker for any error, but I broke it in my recent change in a way that prevents new users from picking a plan in the app

fortunately this hasn't totally broken login because users can still pick a plan in the auth email they get, and probably usually do, but it's a pretty serious regression.

I've tested this pretty well manually, hoping to push asap!