fix: upgrade to enable private spaces (#348)
**Create Space Form**
<img width="998" height="549" alt="image"
src="https://github.com/user-attachments/assets/e5396853-0846-497f-9dde-e2063fd3c923"
/>
Issue: If the user typed anything in the Space name input text, it would
unlock the Private space option.

**Private Spaces**
If the user attempts to access a Private Space directly via URL, it
should redirect him to the Change Plan page - in case the user is not in
a paid plan.

Main Hooks
- `usePrivateSpacesAccess`: is about the user's general ability to use
private spaces
- `useSpaceAccess`: is about access to a specific space, building on top
of the user's general permissions