feat: add blob listing to space detail page
I was trying to track down some issues with deletions in my space yesterday and got far enough into this that I figured I'd polish it off.

This introduces a blob listing page and a separate "blob detail" page that will make it easier for people to track down blobs that aren't part of any particular upload.

Future work:

1) We should make it possible to delete blobs from this page
2) We should add something that filters the list of blobs down to only blobs that don't appear in any existing uploads