fix: better error handling for replica allocate (#373)
When allocating on storage nodes, calling
`allocactionInvocation.execute()` might throw if the HTTP request fails.
In this case a `HandlerExecutionError` will be returned to the agent
calling `space/blob/replicate` and the failure to allocate will not be
recorded. Further the thrown error may prevent other failures/successes
from being recorded and it also means that a retry may select the same
(failed) node(s) which is undesirable.