fix: decrypt delegation should not include plan/get proof  (#375)
The `DecryptDelegation` should not include the `plan/get` proof because
it is not part of the invocation. The `plan/get` delegation must be
passed as part of the `proofs` parameter in the decryption config.
In another PR it will be refactored to attach the `plan/get` delegation
into the `space/content/key/decrypt` invocation.