fix: fixed the bug in @storacha/ui-react
the problem: module resolution/bundling

first off, i don't know how the package had managed to work perfectly fine in the previous versions using tsc to build ui-react. tsc is pretty much the typescript compiler. i don't think it works as a module bundler.

the previous build command in ui-react uses "tsc --build ..." but what it does in essence is transpile @storacha/ui-react and does typechecking where neccessary. it doesn't resolve or bundle any dependencies.

which is why i encountered that "Module not found: Can't resolve '@storacha/ui-core'" error because it basically left the import statement as is without bundling the code and then external projects when they install the dependency run into an issue when that dependency isn't listed in node_modules.

mind you... even when you install @storacha/ui-core separately, the issue doesn't go away.

solution: use esbuild. but with tsup as a wrapper and bundle for the browser platform.

with a tsup config, the ideal thing to do was to just ensure that esbuild bundles the "storacha/ui-core" lib into @storacha/ui-react's build, and now with minification, the bundle size is also reduced.

another thing i observed while trying to fix this for the past couple of days is how, when you install @storacha/ui-react with npm, it fails with a legacy-peer-dependency error and you are encouraged to try installing again by doing the following npm i @storacha/ui-react --legacy-peer-deps.

this is because the latest react version (v19.0.0) in most projects is missing in package.json. i included it.

lastly, i'd recommend that this is tested by everyone locally and i'll await your feedback.