feat: skip index fetch if content link and retrieval auth is provided (#456)
This PR adds an optional content root CID to the `space/index/add`
capability that allows the upload service to `assert/index` on the
indexer without fetching the index.

If a content link and rertieval auth is provided in the invocation then
the upload service can skip fetching the index, delegating that
responsibility to the indexer.

Note: I had to do some spelunking to fix up the ucanto dependencies
(we're depending on a new version).