feat: upgrade to nx 21 and enable OIDC Trusted Publishing (#491)
Also re-enable OIDC Trusted Publishing based package publishing.

This introduces an awkward pre-build step after pnpm install because
that seems to be the only way to get it to build. We should get rid of
it - I'm hoping that continuing the upgrade through nx 21 and next 16
will solve it, but it's nice to have a way to ship this now if we need
to

---------

Co-authored-by: Petra Jaros <peeja@peeja.com>