fix: buffer piece without `insertedAt` date (#540)
We found a piece in the staging pipeline _without_ an `insertedAt` date!

What is also weird is that this piece is the prepend piece.

I did not figure out how this could happen.

This is a temporary fix to the `insertedAt` field issue, but also a fix
to dedupe the prepend pieces from the pieces in the buffer. It's
unlikely this would ever happen but just in case...

<img width="748" height="942" alt="Screenshot 2025-11-18 at 15 49 35"
src="https://github.com/user-attachments/assets/54b7e8b4-3a80-4606-8700-7344fea0ec3c"
/>