feat: cleanup and bugfixing in the signup flow
We need to revive the functionality that stops users from seeing the pricing table after they return to the Storacha auth page. Use the `checkout-success` query param for this rather than the Stripe-style `checkout.session` we were using before.

Also do a bit of cleanup in the private spaces hook to prevent loader jitters on the space creation screen.