feat: add pricing table and plan picker to the CLI (#572)
Take 2 of the plan selection dialog - most of the code is the same as
before, but I added an option to skip it and used it in testing.

I don't currently have a test for `plan setup` but a) it's challenging
test since it involves user input and b) the user can use the web
interface if this is temporarily broken in the future.

I think it might be worth trying to update to a less bespoke testing
framework at some point, and that might make it easier to test user
input with `inquirer`